pe-dashboard-ai50-v3-main/airflow/requirements.txt
feedparser>=6.0.0
requests>=2.31.0
pendulum>=3.0.0
beautifulsoup4>=4.12.0
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
pydantic==2.9.2
requests==2.32.3
python-multipart==0.0.9
openai>=1.35.0,<2
instructor>=1.2.3,<2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-cloud-storage
tabulate
feedparser
chromadb
langchain
langchain-core
langchain-text-splitters
langchain-openai
# apache-airflow==3.1.0
 
 
 